Added mission creation handling

+++ Assets/Scripts/Models/Gun.cs
@@ -3,7 +3,7 @@
 public class Gun {
 	public int id;
 	public string name;
-	public long hitPont	{set; get;}
+	public int hitPont	{set; get;}
 	public int ammo {set; get;}
 	public int maxAmmo {set; get;} // the maximum ammo this gun can have.
 	public int currentAmmo {set; get;}
@@ -11,7 +11,7 @@ public class Gun {
 	public int ammoPrice {set; get;}
 	public int minMedal {set; get;}
 
-	public Gun (int id, string name, long hitPoint, int price, int ammoPrice, int minMedal,  int ammo, int maxAmmo = Constant.maxAmmo) {
+	public Gun (int id, string name, int hitPoint, int price, int ammoPrice, int minMedal,  int ammo, int maxAmmo = Constant.maxAmmo) {
 		this.id = id;
 		this.name = name;
 		this.hitPont = hitPont;
