+++ Assets/Scripts/Models/SpaceShip.cs
@@ -4,14 +4,14 @@
 public class Spaceship {
 	public int id { get; set; }
 	public int price {set; get;}
-	public long hp {set; get;}
+	public int hp {set; get;}
 	public int minMedals {set; get;}
 	public Gun primaryGun {set; get;}
 	public string name { get; set; }
 	public string description {get;set;}
 	public int velocity { get; set;}
 
-	public Spaceship (int id, int price, long hp, int minMedals, Gun primaryGun, string name = "", string description = "") {
+	public Spaceship (int id, int price, int hp, int minMedals, Gun primaryGun, string name = "", string description = "") {
 		this.id = id;
 		this.price = price;
 		this.hp = hp;
